S0C-3A-1A: Add shadow_verify_dual_run_window scenario and update documentation
- Introduced a new scenario: shadow_verify_dual_run_window, which handles dual run window verification.
- Updated the registry to include the new scenario.
- Enhanced documentation to reflect the addition of dual_run_stage2 and its successful validation.
