+++ docs/logs/log-S0C-3A-1A-double-parallel.md
@@ -170,11 +170,27 @@
   - `checks.paging_stability.ok=true`
   - `checks.shared_keys.ok=true`
 
+### Run 7 — dual_run_stage2（OK, 写侧 outbox worker 验证）
+
+- artifact: `summary.json`
+- scenario: `shadow_verify_dual_run_stage2`
+- run_id: `22255653342-1`
+- result: `ok=true`
+- 关键输出（节选）:
+  - `strategy=strict`
+  - `ensure_min_rows=25`, `seed_rows_inserted=25`
+  - `candidate_limit=20`
+  - outbox：`enqueued_total=20`, `done=20`, `failed=0`
+  - worker：`ok=true`, `exit_code=0`, `runtime_seconds≈1.62`
+  - ES：`health_ok=true`, `index_ok=true`, `refresh_ok=true`, `search_ok=true`
+  - `parity_ok=true`
+
 ## Status Update（阶段结论）
 
 - Step B（迁移优先级前三个场景 + readiness gate 复合验证）已通过实证：`Run 1~4` 均 `ok=true`。
 - dual-run（stage1）已迁移并在 GitHub Actions 中跑通：`Run 5` 为 `ok=true`（strict 对齐）。
 - readiness gate 已完成 “shim → scenario” 转接并复跑通过：`Run 6` 为 `ok=true`。
+- dual-run（stage2）已迁移并通过写侧 outbox worker 实证：`Run 7` 为 `ok=true`（strict 对齐）。
 
 
 ## Risks（风险与缓解）
